requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==5.0.2
dnspython==2.8.0
filetype==1.2.0
Flask==3.1.2
google-auth==2.43.0
google-genai==1.54.0
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.1.3
langchain-google-genai==4.0.0
langsmith==0.4.58
MarkupSafe==3.0.3
orjson==3.11.5
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
python-dotenv==1.0.0
PyYAML==6.0.3
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
sqlparse==0.5.4
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uuid_utils==0.12.0
websockets==15.0.1
Werkzeug==3.1.4
whitenoise==6.11.0
zstandard==0.25.0
